Added files via upload
This is my newest version of the web app. It's been renamed for comparison to the other version of the webapp file, due to changes I made in the code.
Still haven't got the piechart working yet, still working on that, but I did manage to get an example of one up.